using float instead of int

+++ main.py
@@ -31,7 +31,7 @@ operations = {
 def calculator():
   print(logo)
   #getting the first number
-  n1 = int(input("What is the first number?: "))
+  n1 = float(input("What is the first number?: "))
 
   #loop through the dictionnary
   for symbol in operations:
@@ -44,7 +44,7 @@ def calculator():
     #operations symbol 
     operation_symbol = input("Pick an operation from the line above:")
     #getting the second number
-    n2 = int(input("What is the second number?: "))
+    n2 = float(input("What is the second number?: "))
     calculation_function = operations[operation_symbol]
     answer = calculation_function(n1, n2)
